reorganize NavigationBlocker; redesign draft dialog; prevent hard navigation
